AtomicList-summarization {IRanges}	R Documentation

AtomicList summarization methods

Description

Operations from the Summary group generic and other summarization operations on AtomicList objects.

Details

The following summarization operations are supported on AtomicList objects:

any, all, min, max, range, sum, prod

mean, var, cov, cor, sd, median, quantile, mad, IQR

which.max, which.min

See S4groupGeneric for more details.

Note that the which.min and which.max functions have an extra argument, global=FALSE, which controls whether the returned subscripts are global (compatible with the unlisted form of the input) or local (compatible with the corresponding list element).

See Also

AtomicList_utils for common operations on AtomicList objects.

AtomicList objects.

Examples

x1 <- c(1L, 2L, 3L, 5L, 2L, 8L)
x2 <- c(15L, -5L, 80L, -1L, 15L, 79L, 20L, 5L)
il <- IntegerList(one=x1, x2)

sum(il)        # equivalent to 'sapply(il, sum)'
mean(il)       # equivalent to 'sapply(il, mean)'
range(il)      # equivalent to 'sapply(il, range)'
which.max(il)  # equivalent to 'sapply(il, which.max)'

[Package IRanges version 2.44.0 Index]