DataFrame-utils {IRanges}	R Documentation

Common operations on DataFrame objects

Description

Common operations on DataFrame objects.

Splitting

In the following code snippets, x is a DataFrame.

split(x, f, drop = FALSE): Splits x into a CompressedSplitDataFrameList, according to f, dropping elements corresponding to unrepresented levels if drop is TRUE.

mstack(..., .index.var = "name"): Stacks the data frames passed as through ..., using .index.var as the index column name. See stack.

Author(s)

Michael Lawrence

See Also

DataTable and Vector

Examples


## split

sw <- DataFrame(swiss)
swsplit <- split(sw, sw[["Education"]])
  
## rbind

do.call(rbind, as.list(swsplit))

## cbind

cbind(DataFrame(score), DataFrame(counts))

[Package IRanges version 2.16.0 Index]