List-class-leftovers {IRanges}	R Documentation

List objects (old man page)

Description

IMPORTANT NOTE - 9/4/2014: This man page is being refactored. Most of the things that used to be documented here have been moved to the man page for List objects located in the S4Vectors package.

Details

The only thing left here is the documentation of the stack method for List objects. In the code snippets below, x is a List object.

stack(x, index.var = "name", value.var = "value"): As with stack on a list, constructs a DataFrame with two columns: one for the unlisted values, the other indicating the name of the element from which each value was obtained. index.var specifies the column name for the index (source name) column and value.var specifies the column name for the values.

See Also

The List class defined and documented in the S4Vectors package.

Examples


starts <- IntegerList(c(1, 5), c(2, 8))
ends <- IntegerList(c(3, 8), c(5, 9))
rgl <- IRangesList(start=starts, end=ends)
rangeDataFrame <- stack(rgl, "space", "ranges")

[Package IRanges version 2.16.0 Index]