ViewsList-class {IRanges}	R Documentation

List of Views

Description

An extension of List that holds only Views objects.

Details

ViewsList is a virtual class. Specialized subclasses like e.g. RleViewsList are useful for storing coverage vectors over a set of spaces (e.g. chromosomes), each of which requires a separate RleViews object.

As a List subclass, ViewsList inherits all the methods available for List objects. It also presents an API that is very similar to that of Views, where operations are vectorized over the elements and generally return lists.

Author(s)

P. Aboyoun and H. Pagès

See Also

List-class, RleViewsList-class.

findOverlaps.

Examples


  showClass("ViewsList")

[Package IRanges version 2.16.0 Index]