seqapply {IRanges}	R Documentation

2 methods that should be documented somewhere else

Description

unsplit method for List object and split<- method for Vector object.

Usage

## S4 method for signature 'List'
unsplit(value, f, drop = FALSE)
## S4 replacement method for signature 'Vector'
split(x, f, drop = FALSE, ...) <- value

Arguments

value
The List object to unsplit.

f
A factor or list of factors

drop
Whether to drop empty elements from the returned list

x
Like X

...
Extra arguments to pass to FUN

Details

unsplit unlists value, where the order of the returned vector is as if value were originally created by splitting that vector on the factor f.

split(x, f, drop = FALSE) <- value: Virtually splits x by the factor f, replaces the elements of the resulting list with the elements from the list value, and restores x to its original form. Note that this works for any Vector, even though split itself is not universally supported.

Author(s)

Michael Lawrence

[Package IRanges version 2.44.0 Index]